% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_sign}
\alias{add_sign}
\title{Add signs to variable names}
\usage{
add_sign(names, signs)
}
\arguments{
\item{names}{vector of variable names.}

\item{signs}{vector of signs (the same length as names).}
}
\description{
Add signs to variable names
}
