% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{boot_ci}
\alias{boot_ci}
\title{Bootstrap Confidence Intervals}
\usage{
boot_ci(boot.res, alpha = 0.05)
}
\arguments{
\item{boot.res}{matrix of bootstrap replicates}

\item{alpha}{significance level}
}
\value{
matrix of confidence intervals
}
\description{
Calculates bootstrap confidence intervals for matrix of bootstrap replicates
}
