% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{boot_default}
\alias{boot_default}
\title{Bootstrap (default)
Bootstrap function for the non-parametric and the nearest neighbor methods}
\usage{
boot_default(func, Y, Y_pos, X, X_std, control, wgt, n.quant, lambda,
  sigma, grp.size, n.boot, quick)
}
\value{
a list - the output from the function 'boot()'.
}
\description{
Bootstrap (default)
Bootstrap function for the non-parametric and the nearest neighbor methods
}
