% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dev_moments}
\alias{dev_moments}
\title{Moment deviation}
\usage{
dev_moments(beta, base, controls, wgt, ...)
}
\arguments{
\item{beta}{a lagrange multiplier}

\item{base}{basline weights}

\item{controls}{controls matrix (with a constant)}

\item{wgt}{original weights}
}
\value{
vector of moment deviations
}
\description{
Finds the moment deviation for a given lagrange multiplier
}
