% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{kl_dist.def}
\alias{kl_dist.def}
\title{Kullback-Leibler Divergence}
\usage{
kl_dist.def(wgt, wgt1, ...)
}
\arguments{
\item{wgt}{original weights.}

\item{wgt1}{weights under I = 1.}
}
\value{
scalar of kullback-liebler divergence.
}
\description{
Calculates Kullback-Liebler Divergence for two weight vectors.
}
