% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{kl_dist_cor}
\alias{kl_dist_cor}
\title{Kullback-Leibler Divergence}
\usage{
kl_dist_cor(X, wgt, ni)
}
\arguments{
\item{X}{numeric data frame or matrix of factors to be considered.}

\item{wgt}{an optional vector of weights.}

\item{ni}{difference in means (mu1 - mu0)}
}
\value{
scalar of kullback-liebler divergence.
}
\description{
Calculates Kullback-Liebler Divergence for two multivariate normal distributions.
}
