% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nn}
\alias{nn}
\title{Nearest-neighbors method}
\usage{
nn(Y, X, controls = NULL, wgt = rep(1, length(Y)), lambda = 100,
  sigma = 1, ...)
}
\arguments{
\item{Y}{outcome vector (must be numeric without NA's).}

\item{X}{data frame or matrix of factors to be considered.}

\item{controls}{data frame or matrix of factors to control for. these are factors
that we can't consider while looking for the optimal intervention
(e.g. gender).}

\item{wgt}{an optional vector of weights.}

\item{lambda}{the lagrange multiplier. also known as the shadow price of an
intervention.}

\item{sigma}{distance penalty for the nearest-neighbors method.}
}
\value{
vector of weights under I = 1
}
\description{
Calculates weights under I = 1, using the nearest-neighbors method
}
