% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{non_parm}
\alias{non_parm}
\title{Non-parametric method}
\usage{
non_parm(Y, X, controls, wgt, lambda)
}
\arguments{
\item{Y}{outcome}
}
\value{
vector of weights under I = 1
}
\description{
Calculates weights under I = 1, using the non-parametric method
}
