% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{par_cor}
\alias{par_cor}
\title{Partial Correlation}
\usage{
par_cor(Y, X, control = NULL, wgt = rep(1, length(Y)), ...)
}
\arguments{
\item{Y}{outcome vector (must be numeric without NA's).}

\item{X}{data frame or matrix of factors to be considered.}

\item{control}{data frame or matrix of factors to control for. these are factors
that we can't consider while looking for the optimal intervention
(e.g. race).}

\item{wgt}{an optional vector of weights.}
}
\value{
data frame with partial correlations & covariance
}
\description{
Calculates correlation / covariance between Y and X, holding control constant
}
