% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.optint}
\alias{plot.optint}
\title{Plot optint object}
\usage{
\method{plot}{optint}(object, plot.vars = "all", plot.ci = T,
  .col = 1, ...)
}
\arguments{
\item{object}{an optint object.}

\item{plot.vars}{which variables to plot? "all" (default) plot all variables and
"sig" plot only siginficance variables. Another option is vector
 with variables names to plot.}

\item{plot.ci}{logical. if TRUE (default) plot condifendece intervals. Otherwise
plot only point estimates.}
}
\description{
Produce variables important plot from an optint object.
}
