% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.optint}
\alias{summary.optint}
\title{Summary for optint object}
\usage{
\method{summary}{optint}(object, r = 4)
}
\arguments{
\item{object}{an optint object.}

\item{r}{number of decimal places to use.}
}
\description{
Report results from an optint object.
}
