% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{wgt_adjust}
\alias{wgt_adjust}
\title{Weights adjustment}
\usage{
wgt_adjust(control, base.wgt1, wgt)
}
\arguments{
\item{control}{numeric data frame or matrix of factors to control for. these are factors
that we can't consider while looking for the optimal intervention
(e.g. race).}

\item{base.wgt1}{baseline weights under I = 1}

\item{wgt}{an optional vector of weights.}
}
\value{
vector of adjusted weights under I = 1
}
\description{
Adjust new weights so that the distribution of the control variables would not change
}
