% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{wtd_bin}
\alias{wtd_bin}
\title{Weighted Bin}
\usage{
wtd_bin(x, n.quant, wgt)
}
\arguments{
\item{x}{variable.}

\item{n.quant}{number of quantiles.}

\item{wgt}{vector of weigts.}
}
\value{
vector of means.
}
\description{
Divide the data into equal size bins and calculate mean
}
