% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optint.R
\name{optint}
\alias{optint}
\title{Optimal intervention}
\usage{
optint(Y, X, controls = NULL, wgt = rep(1, length(Y)),
  method = "non-parametric", lambda = 100, sigma = 1,
  grp.size = 30, n.boot = 1000, sign.factor = 2/3, alpha = 0.05,
  n.quant = length(Y)/10, seed = runif(1, 0, .Machine$integer.max),
  ...)
}
\arguments{
\item{Y}{outcome vector (must be numeric without NA's).}

\item{X}{data frame or matrix of factors to be considered.}

\item{controls}{data frame or matrix of factors to control for. these are factors
that we can't consider while looking for the optimal intervention
(e.g. race).}

\item{wgt}{an optional vector of weights.}

\item{method}{the method to be used. one of "correlation" / "non-parametric" /
"nearest-neighbors".}

\item{lambda}{the lagrange multiplier. also known as the shadow price of an
intervention.}

\item{sigma}{distance penalty for the nearest-neighbors method.}

\item{n.boot}{number of bootstrap replications to use for the standard errors
calculation.}

\item{n.quant}{}

\item{seed}{}
}
\value{
an optint object.
}
\description{
identifies the factors with the greatest potential to increase
a pre-specified outcome, using varius methods.
}
