% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optint.R
\name{optint_by_group}
\alias{optint_by_group}
\title{Optimal intervention, by group}
\usage{
optint_by_group(Y, X, group, control = NULL, wgt = rep(1, length(Y)),
  method = "non-parametric", lambda = 100, sigma = 1,
  grp.size = 30, n.boot = 1000, alpha = 0.05)
}
\arguments{
\item{Y}{outcome vector (must be numeric without NA's).}

\item{X}{numeric data frame or matrix of factors to be considered.}

\item{group}{vector with group labels (i.e. grouping variable). the function
\code{\link[optinterv]{optint}} implemented for each group separately.}

\item{control}{numeric data frame or matrix of factors to control for. these are factors
that we can't consider while looking for the optimal intervention
(e.g. race).}

\item{wgt}{an optional vector of weights.}

\item{method}{the method to be used. either "non-parametric" (default), "correlation" or
"nearest-neighbors".}

\item{lambda}{the lagrange multiplier. also known as the shadow price of an
intervention.}

\item{sigma}{distance penalty for the nearest-neighbors method.}

\item{grp.size}{for the nearest-neighbors method; if the number of examples in each
control group is smaller than grp.size, performs weight adjustment
using wgt_adjust().else, calculate weights seperatly for each
control group.}

\item{n.boot}{number of bootstrap replications to use for the standard errors /
confidence intervals calculation.}

\item{alpha}{significance level for the confidence intervals.}
}
\description{
Similar to \code{\link[optinterv]{optint}}, identifies the factors with the greatest
potential to increase a pre-specified outcome for each group separately, and thus allowing
to detect heterogeneity between groups.
}
