% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{perm_test}
\alias{perm_test}
\title{Permutation test
Test the null hypothesis P(X|I=0) = P(X|I=1), using permutation test.}
\usage{
perm_test(estimates, wgt, wgt1, X, n.quant, n.perm = 1000, ...)
}
\arguments{
\item{estimates}{point estimates of the percentile distance between P(X|I=0) & P(X|I=1).}

\item{wgt}{an optional vector of weights.}

\item{X}{numeric data frame or matrix of factors to be considered.}

\item{n.quant}{number of quantiles to use when calculating CDF distance.}

\item{n.perm}{number of permutations to permute from wgt1.}
}
\value{
vector of p values.
}
\description{
Permutation test
Test the null hypothesis P(X|I=0) = P(X|I=1), using permutation test.
}
