% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.optint}
\alias{plot.optint}
\title{Plot optint object}
\usage{
\method{plot}{optint}(object, plot.vars = "sig", plot.ci = T,
  graph.col = 1, alpha = 0.05, ...)
}
\arguments{
\item{object}{an optint object.}

\item{plot.vars}{which variables to plot? either a number - indicating to plot the first n variables,
"sig" (default) - plot only significant  variables, or a vector
 with names of variables to plot.}

\item{plot.ci}{logical. if TRUE (default) plot confidence intervals. Otherwise
plot only point estimates.}

\item{graph.col}{graph color/s.}

\item{alpha}{significance level for the confidence intervals. also
used in order to determine which variables are significant.}
}
\description{
Produce variables important plot from an optint object.
}
