% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.optint}
\alias{plot.optint}
\title{Plot optint object}
\usage{
\method{plot}{optint}(object, plot.vars = "sig", plot.ci = T,
  graph.col = 1, alpha = 0.05, ...)
}
\arguments{
\item{object}{an optint object.}

\item{plot.vars}{which variables to plot? "all" plot all variables and
"sig" (default) plot only siginficance variables. Another option is a vector
 with names of variables to plot.}

\item{plot.ci}{logical. if TRUE (default) plot condifendece intervals. Otherwise
plot only point estimates.}

\item{graph.col}{graph color/s.}

\item{alpha}{significance level. only used if plot.vars = "sig".}
}
\description{
Produce variables important plot from an optint object.
}
