% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.optint_by_group}
\alias{plot.optint_by_group}
\title{Plot optint object, by group}
\usage{
\method{plot}{optint_by_group}(object, plot.vars = "sig",
  graph.col = NULL, alpha = 0.05, ...)
}
\arguments{
\item{object}{an optint_by_group object.}

\item{plot.vars}{which variables to plot? either a number - indicating to plot the first n variables,
"sig" (default) - plot only significant  variables, or a vector
 with names of variables to plot.}

\item{graph.col}{graph color/s.}

\item{alpha}{significance level for the confidence intervals. also
used in order to determine which variables are significant.}
}
\description{
Produce variables important plot from an optint_by_group object.
}
