% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_change}
\alias{plot_change}
\title{Plot the change in the distribution of X}
\usage{
plot_change(object, plot.vars = "sig", graph.col = c("lightcoral",
  "cadetblue3"), alpha = 0.05, print.sep = T, line.type = c(1, 2),
  ...)
}
\arguments{
\item{object}{an optint object.}

\item{plot.vars}{which variables to plot? either a number (n) -
indicating to plot the first n variables,
"sig" (default) - plot only significant  variables, or a vector
 with names of variables to plot.}

\item{graph.col}{graph color/s.}

\item{alpha}{significance level for the confidence intervals. also
used in order to determine which variables are significant.}

\item{print.sep}{logical. If TRUE (default) plot each graph seperatly.
This option is highly recommended for more than 4 variables.}
}
\description{
Plot denisty or barchart of X, before and after the intervention.
}
