% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{var_pos}
\alias{var_pos}
\title{Variable Position}
\usage{
var_pos(object, plot.vars = "sig", alpha, ...)
}
\arguments{
\item{object}{an optint object.}

\item{plot.vars}{which variables to plot? "all" plot all variables and
"sig" (default) plot only siginficance variables. Another option is a vector
 with names of variables to plot.}

\item{alpha}{significance level. only used if plot.vars = "sig".}
}
\value{
vector of variables incidents
}
\description{
Find which variables to plot.
}
